Moved RocksDB related checkpoint data structure to RocksDBCheckpointUtils.h
